Add startup-beep
This is a very helpful startup script for headless machines.

The machine will always beep once when it boots up.
It will then try to ping the gateway (a test for a network connection),
and will beep again if it can.

If your machine doesn't beep twice, you can't ssh into it.