Dracut USE: fix when package.use does not exist
This section of code checks to see if any dracut modules need
USE=device-mapper support (by default, we do not provide it).

However, the code did not do its job when /etc/portage/package.use
does not exist.

In the code path that follows the package.use directory, when package.use
does not exist, create it, then write the USE flag.