Fix fstype detection
This would break when fs labels were used as the field numbering would
change. Make the awks detect where the TYPE= field is, as it is
required.

Also, stop using IFS to break up the mount command as it's pretty buggy
and hard to test. Don't use awk to make the $INSTDIR substitution for
the same reason.